backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9,<2.10
django-cors-headers>=4.4,<4.5
django-filter>=24.3,<25.0
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
